Implement prerelease target
Going to need some kind of shim in order to use this, though.

Currently, the code will end up trying to call `System.Version.Parse` on
strings of the form "1.2.3-alpha", and the "-alpha" will make it
explode.

My plan is to add a shim here that will strip the pre/meta parts out so
that this can be used in the time being, to be removed when the
appropriate fix is made in FSharp.Version.Utils